#341: Move new helper methods from `core` to `utils` and rename them according to the naming convention
