+++ dataclasses_json/core.py
@@ -12,7 +12,7 @@ from datetime import datetime, timezone
 from decimal import Decimal
 from enum import Enum
 from typing import (Any, Collection, Mapping, Union, get_type_hints,
-                    Tuple, Type, TypeVar)
+                    Tuple, TypeVar)
 from uuid import UUID
 
 from typing_inspect import is_union_type  # type: ignore
@@ -22,6 +22,9 @@ from dataclasses_json.utils import (_get_type_cons, _get_type_origin,
                                     _handle_undefined_parameters_safe,
                                     _is_collection, _is_mapping, _is_new_type,
                                     _is_optional, _isinstance_safe,
+                                    _get_type_arg_param,
+                                    _get_type_args,
+                                    _NO_ARGS,
                                     _issubclass_safe)
 
 Json = Union[dict, list, str, int, float, bool, None]
@@ -86,42 +89,6 @@ def _user_overrides_or_exts(cls):
         field_config.update(field.metadata.get('dataclasses_json', {}))
         overrides[field.name] = FieldOverride(*map(field_config.get, confs))
     return overrides
-
-
-class _NoArgs(object):
-    def __bool__(self):
-        return False
-
-    def __len__(self):
-        return 0
-
-    def __iter__(self):
-        return self
-
-    def __next__(self):
-        raise StopIteration
-
-
-_NO_ARGS = _NoArgs()
-
-
-def _extract_args(tp: Type, default: Tuple[Type, ...] = _NO_ARGS) -> \
-        Union[Tuple[Type, ...], _NoArgs]:
-    if hasattr(tp, '__args__'):
-        if tp.__args__ is not None:
-            return tp.__args__
-    return default
-
-
-def _extract_type_parameter(tp: Type, index: int) -> Union[Type, _NoArgs]:
-    _args = _extract_args(tp)
-    if _args is not _NO_ARGS:
-        try:
-            return _args[index]
-        except (TypeError, IndexError, NotImplementedError):
-            pass
-
-    return _NO_ARGS
 
 
 def _encode_json_type(value, default=_ExtendedEncoder().default):
@@ -287,14 +254,14 @@ def _decode_generic(type_, value, infer_missing):
     # FIXME this is a hack to fix a deeper underlying issue. A refactor is due.
     elif _is_collection(type_):
         if _is_mapping(type_):
-            k_type, v_type = _extract_args(type_, (Any, Any))
+            k_type, v_type = _get_type_args(type_, (Any, Any))
             # a mapping type has `.keys()` and `.values()`
             # (see collections.abc)
             ks = _decode_dict_keys(k_type, value.keys(), infer_missing)
             vs = _decode_items(v_type, value.values(), infer_missing)
             xs = zip(ks, vs)
         else:
-            xs = _decode_items(_extract_type_parameter(type_, 0),
+            xs = _decode_items(_get_type_arg_param(type_, 0),
                                value, infer_missing)
 
         # get the constructor if using corresponding generic type in `typing`
@@ -304,12 +271,12 @@ def _decode_generic(type_, value, infer_missing):
         except (TypeError, AttributeError):
             res = type_(xs)
     else:  # Optional or Union
-        _args = _extract_args(type_)
+        _args = _get_type_args(type_)
         if _args is _NO_ARGS:
             # Any, just accept
             res = value
         elif _is_optional(type_) and len(_args) == 2:  # Optional
-            type_arg = _extract_type_parameter(type_, 0)
+            type_arg = _get_type_arg_param(type_, 0)
             if is_dataclass(type_arg) or is_dataclass(value):
                 res = _decode_dataclass(type_arg, value, infer_missing)
             elif _is_supported_generic(type_arg):
